requirements.txt
requests
pymongo
flask
Flask-PyMongo
bs4
python-dotenv
uvicorn
fastapi
jinja2
aiofiles
sqlitedict
natsort